cmd/dist: rationalize test names
This changes all dist test names to be either of the form
go_test(_bench):<pkg> for std and cmd packages, or <pkg>[:<variant>]
for tests that are not part of the default std cmd set. (Ideally we
would make everything <pkg>[:<variant>], but the go_test: prefix is
known to the build coordinator.)

The plan is to rewrite the Package field of the test JSON output to
always use <pkg>[:<variant>] so that we can disambiguate output from
differently configured runs of the same test package, and naturally
flow this through to any test result viewer.

For #37486.

Change-Id: I91e66c1eb8b97345a37b3ee6c44487abcfeca787