cmd/dist: introduce test variants and rationalize test names
This introduces the concept of test variants in dist, which are
different configurations of the same package. The variant of a test is
a short string summarizing the configuration.

Currently, all dist tests have a unique name. Using the variant, we
can normalize these names so that all dist test names to be either of
the form go_test(_bench):<pkg> for std and cmd packages, or
<pkg>[:<variant>] for tests that are not part of the default std cmd
set. (Ideally we would make everything <pkg>[:<variant>], but the
 go_test: prefix is known to the build coordinator.)

The plan is to rewrite the Package field of the test JSON output to
always use <pkg>[:<variant>] so that we can disambiguate output from
differently configured runs of the same test package, and naturally
flow this through to any test result viewer.

For #37486.

Change-Id: I17cc32d662c4678b8cf4f5c5f89f81497ec43f2e