runtime: derive trace event from wait reason
Currently the trace event type gets plumbed down through gopark
manually, but we have enough unique wait reasons (which also must get
plumbed down) that we can easily derive the right trace event to emit
from just the wait reason.

There are a few wait reasons that don't apply to the tracer, so for
these cases we skip emitting an event. If there is a bug, this will
likely create an invalid trace, though for all existing waitReasons it
doesn't actually make sense to emit an event anyway.

An alternative is to crash the program, but that seems extreme for what
is a diagnostics issue. However, to help debug such issues, this change
also adds a debug mode that does crash the program if an invalid
waitReason does make its way into traceGoPark.

The goal of this change is to hide more of the tracer's details inside
of trace.go, in this case the traceEv* values specifically.

Change-Id: Id633e1704d2c8838c6abd1214d9695537c4ac7db