cmd/go/internal/test: add -label flag
When running go test with various different settings, it can be useful
to distinguish those results from others in the test output. This CL
adds a new go test flag -label=arg that adds :arg to the printed
package names in test results, so

% GODEBUG=cpu/sse2=off go test -label=sse2off strings math
ok  	strings:sse2off	0.886s
ok  	math:sse2off	0.138s
%

This label is also included in the Package field in go test -json
output. This is particularly useful when running a test in different
modes in a larger wrapper - it avoids the need to post-process the
JSON output to adjust the names to keep different modes separated.

Fixes #60128
For #37486

Change-Id: I9a8ecbf7d890ef4079bc0179203b2d5327f370a6