cmd/gofmt: don't clobber original file on permission error
When gofmt needs to rewrite a file, it first copies it into a backup.
If the rewrite fails, it renames the backup to the original.
However, if for some reason the file is owned by some other user,
and if the rewrite fails because gofmt doesn't have permission to
write to the file, then renaming the backup file will change
the file owner. This CL changes gofmt so that if it gets a permission
error writing the file, it assumes that the file is unchanged,
and it does not rename the backup file to clobber the original.

Also create the backup file with the correct permissions,
to avoid a tiny gap when some process might get write access to the
file contents that it shouldn't have. (This tiny gap only applies to
files that are not formatted correctly, and have read-only permission,
and are in a directory with write permission.)

Fixes #60225

Change-Id: Ic16dd0c85cf416d6b2345e0650d5e64413360847