cmd/compile: permanently enable more lenient type inference for untyped arguments
This CL removes the ability to switch back and forth between original
and new behavior. The new behavior will be available to all Go code,
even when -lang < go1.21: old code will keep working as before, but
old code that didn't compile may now compile.

To be submitted once #58671 is accepted.

For #58671.

Change-Id: I83a1d393f0ce7871be8f38ec35742d393946c55f