cmd/internal/obj: move arch specific code out of writeDconv
The assembler uses function Rconv to write a register into string form
when printing instructions. Rconv is specified by every architecture so
we can use same code to handle all the registers that come from
different arches. However, there are still some special cases that we
need to hard code, which is not elegant and hard to maintain.

This CL moves the arch specific code out of writeDconv by adding a new
list of print functions, each of the functions will be used to
print a unique type of register for the specific arch. After that, we
can easily handle the special cases in the backend of every arch.

Change-Id: I129e9c0a32993d945878b09f6379087ef845902d