cmd/compile: report an error URL with error messages
Add Config.ErrorURL (or Config._ErrorURL for go/types) to configure
whether and how an error message should report a URL for errors that
have an error code.

In the compiler, configure types2 to report an error URL of the form
" [go.dev/e/XXX]", where XXX stands for the error code, with the URL
appended to the first line of an error.

Example for a one-line error message:

<pos>: undefined: f [go.dev/e/UndeclaredName]

Example for a multi-line error message:

<pos>: not enough arguments in call to min [go.dev/e/WrongArgCount]
	have ()
	want (P, P)

Change-Id: I26651ce2c92ad32fddd641f003db37fe12fdb1cd